requirements.txt
Django>=2.0.0, < 2.2.13 
mysqlclient>=2.0.1
pytz==2018.5
whitenoise==4.1
tablib==0.14.0
django-admin-menu==1.2.1
django-import-export
djangorestframework==3.9.2
djangorestframework-simplejwt==4.4.0
git+git://github.com/ahoazure/django-admin-menu.git
git+git://github.com/ahoazure/django-data-wizard.git
Pillow>=6.0.0
django-admin-list-filter-dropdown==1.0.2
django-rest-swagger>=2.2.0
django-parler==2.0.1
django-parler-rest==2.1
django-modeladmin-reorder==0.3.1